api/requirements.txt
flask
flask_cors
python-dotenv
langchain
langchain-openai
langchain-core
pydantic